Confirm credential after warning dialog when installing CA certificate
This is part of the changes to improve the UX and language for installing certificates.
Previously, confirm credentials occurs once a CA certificate is selected
from the file picker. This CL moves the confirm credential logic to appear after the
warning dialog and before the file picker, when installing CA certificates.

Bug: 139173976
Test: manual testing from Settings by selecting the certificate type
	preference and ensuring the installation flow still worked as expected.

Change-Id: Iba737f391fa672a305daae475467d5d25a6206df