Created dialogs when installing certificates outside Settings
This is part of the changes to improve the UX and language for installing certificates.
Previously, all certificate types could be installed from the Settings app and other apps.
This CL prevents CA certificates from being installed outside of the Settings app.
It also introduces new dialogs:
 - CA certificate: dialog to redirect user to Settings
 - User/WiFi certificate: dialog to select certificate usage and enter certificate name

Bug: 139173976
Test: manual testing by installing different certificate types from the Files app
      Including testing certificate which needs to be extracted
      go/enterprise-wifi-lab-access

Screenshot of the screen: https://hsv.googleplex.com/5073669078384640
https://hsv.googleplex.com/5697514017652736

Change-Id: I1a456ef9d8f139345b6159e1ad83d2d3a14687f8