Added invalid certificate dialog
This is part of the changes to improve the UX and language for installing certificates.
With the addition of the Settings page to select the type of certificate to install,
additional logic is required to ensure an appropriate certificate is selected.
This CL does validation checks and will show a dialog if the wrong certificate type is selected.

Bug: 139173976
Test: manual testing by selecting different certificate types with the file picker
      Including testing certificate which needs to be extracted
      go/enterprise-wifi-lab-access

Screenshot of the screen: https://hsv.googleplex.com/4890125932691456

Change-Id: I45064f7b5a14d84232d7009fa7c2b1387f674ae4